+++ models/datasource.js
@@ -16,8 +16,8 @@ teaDataSource
   })
 
   .catch((err) => {
-    console.log(err);
-    console.log("Failed to connect database");
+    console.log("Failed to connect database", err);
+    teaDataSource.destroy();
   });
 
 module.exports = {
